Downloads the podcast episodes from https://www.unlimitedspanish.com/podcasts/ - the excellent Spanish learning resource.
It parses each episode page and obtains the URLs of .mp3 and .pdf files as well as the episode name.
Then it downloads these files and renames it to the episode name.
Different formats of URLs and names for different episode diapasons as well as known special cases (typos and not only those) for some episodes are taken into account.

TODO: remove debug information and make it possible to take episode numbers and diapasons from the command line and check if there are known special cases for episodes after 80.